fix : 무한 렌더링 문제
